:construction: WIP
TRAV-33

The menu item is an important part of both main and submenu. We need to develop
this component.

On this commit, we're going to develop the menu item component to use on any
menu we need on the website.

Now we have a component that can be used on all our websites.