:sparkles: Adds internacionalizaton features
TRAV-37

After creating the basic structure for the TopMenu component, we need now to
code the basic rules for it.

On this commit I've created a basic logic to display a different message
depending on the website the user is currently on.

Now, we can proceed to coding the social network icons.